# refactor(rust): Dispatch new-streaming NDJSON source to updated multiscan