--- a/crates/polars-stream/src/nodes/io_sources/multi_file_reader/initialization/mod.rs
+++ b/crates/polars-stream/src/nodes/io_sources/multi_file_reader/initialization/mod.rs
@@ -1,4 +1,5 @@
 pub mod predicate;
+pub mod projection;
 pub mod slice;
 
 use std::sync::{Arc, Mutex};