--- a/crates/polars-stream/src/nodes/io_sources/multi_file_reader/initialization/slice.rs
+++ b/crates/polars-stream/src/nodes/io_sources/multi_file_reader/initialization/slice.rs
@@ -97,7 +97,11 @@ async fn resolve_negative_slice(config: &MultiFileReaderConfig) -> PolarsResult<
                         .unwrap()
                         .into_owned()
                         .map(|source| {
-                            file_reader_builder.build_file_reader(source, cloud_options.clone())
+                            file_reader_builder.build_file_reader(
+                                source,
+                                cloud_options.clone(),
+                                scan_source_idx,
+                            )
                         })?;
 
                 if verbose {