--- a/crates/polars-stream/src/nodes/io_sources/multi_file_reader/reader_interface/builder.rs
+++ b/crates/polars-stream/src/nodes/io_sources/multi_file_reader/reader_interface/builder.rs
@@ -18,5 +18,6 @@ pub trait FileReaderBuilder: Debug + Send + Sync + 'static {
         &self,
         source: ScanSource,
         cloud_options: Option<Arc<CloudOptions>>,
+        scan_source_idx: usize,
     ) -> Box<dyn FileReader>;
 }