--- a/crates/polars-stream/src/nodes/io_sources/ndjson/chunk_reader.rs
+++ b/crates/polars-stream/src/nodes/io_sources/ndjson/chunk_reader.rs
@@ -1,10 +1,8 @@
-use std::sync::Arc;
-
-use polars_core::schema::{SchemaExt, SchemaRef};
+use polars_core::schema::SchemaRef;
 use polars_error::PolarsResult;
+use polars_io::ndjson;
 use polars_io::prelude::parse_ndjson;
 use polars_plan::dsl::NDJsonReadOptions;
-use polars_utils::pl_str::PlSmallStr;
 
 use crate::nodes::compute_node_prelude::*;
 
@@ -20,18 +18,9 @@ pub(super) struct ChunkReader {
 impl ChunkReader {
     pub(super) fn try_new(
         options: &NDJsonReadOptions,
-        reader_schema: &SchemaRef,
-        with_columns: Option<&[PlSmallStr]>,
+        projected_schema: &SchemaRef,
     ) -> PolarsResult<Self> {
-        let projected_schema: SchemaRef = if let Some(cols) = with_columns {
-            Arc::new(
-                cols.iter()
-                    .map(|x| reader_schema.try_get_field(x))
-                    .collect::<PolarsResult<_>>()?,
-            )
-        } else {
-            reader_schema.clone()
-        };
+        let projected_schema = projected_schema.clone();
 
         #[cfg(feature = "dtype-categorical")]
         let _cat_lock = projected_schema
@@ -48,6 +37,10 @@ impl ChunkReader {
     }
 
     pub(super) fn read_chunk(&self, chunk: &[u8]) -> PolarsResult<DataFrame> {
-        parse_ndjson(chunk, None, &self.projected_schema, self.ignore_errors)
+        if self.projected_schema.is_empty() {
+            Ok(DataFrame::empty_with_height(ndjson::count_rows(chunk)))
+        } else {
+            parse_ndjson(chunk, None, &self.projected_schema, self.ignore_errors)
+        }
     }
 }