--- a/crates/polars-stream/src/nodes/io_sources/ndjson/row_index_limit_pass.rs
+++ b/crates/polars-stream/src/nodes/io_sources/ndjson/row_index_limit_pass.rs
@@ -6,17 +6,13 @@ use polars_io::RowIndex;
 use polars_utils::IdxSize;
 use polars_utils::priority::Priority;
 
-use crate::async_executor::{AbortOnDropHandle, TaskPriority, spawn};
-use crate::async_primitives::distributor_channel::distributor_channel;
 use crate::async_primitives::linearizer::Linearizer;
-use crate::async_primitives::wait_group::WaitGroup;
-use crate::async_primitives::{connector, distributor_channel};
-use crate::morsel::{Morsel, MorselSeq};
-use crate::nodes::io_sources::MorselOutput;
+use crate::morsel::{Morsel, MorselSeq, SourceToken};
+use crate::nodes::io_sources::multi_file_reader::reader_interface::output::FileReaderOutputSend;
 
 pub struct ApplyRowIndexOrLimit {
     pub morsel_receiver: Linearizer<Priority<Reverse<MorselSeq>, DataFrame>>,
-    pub phase_tx_receivers: Vec<connector::Receiver<MorselOutput>>,
+    pub morsel_tx: FileReaderOutputSend,
     pub limit: Option<usize>,
     pub row_index: Option<RowIndex>,
     pub verbose: bool,
@@ -26,7 +22,7 @@ impl ApplyRowIndexOrLimit {
     pub async fn run(self) -> PolarsResult<()> {
         let ApplyRowIndexOrLimit {
             mut morsel_receiver,
-            phase_tx_receivers,
+            mut morsel_tx,
             limit,
             row_index,
             verbose,
@@ -42,9 +38,6 @@ impl ApplyRowIndexOrLimit {
                 &limit, &row_index
             );
         }
-
-        let (mut morsel_distributor, phase_sender_handles) =
-            init_morsel_distributor(phase_tx_receivers);
 
         let mut n_rows_received: usize = 0;
 
@@ -74,7 +67,13 @@ impl ApplyRowIndexOrLimit {
 
             n_rows_received = n_rows_received.saturating_add(df.height());
 
-            if morsel_distributor.send((morsel_seq, df)).await.is_err() {
+            // No wait group logic here, already attached by line batch processors.
+
+            if morsel_tx
+                .send_morsel(Morsel::new(df, morsel_seq, SourceToken::new()))
+                .await
+                .is_err()
+            {
                 break;
             }
 
@@ -85,15 +84,6 @@ impl ApplyRowIndexOrLimit {
 
         // Explicit drop to stop NDJSON parsing as soon as possible.
         drop(morsel_receiver);
-        drop(morsel_distributor);
-
-        if verbose {
-            eprintln!("[NDJSON ApplyRowIndexOrLimit]: wait for morsel distributor handles");
-        }
-
-        for handle in phase_sender_handles {
-            handle.await?;
-        }
 
         if verbose {
             eprintln!("[NDJSON ApplyRowIndexOrLimit]: returning");
@@ -101,62 +91,4 @@ impl ApplyRowIndexOrLimit {
 
         Ok(())
     }
-}
-
-/// Initialize and connect a distributor to the morsel outputs. Returns the distributor and the join
-/// handles for the tasks that send to the morsel outputs.
-#[allow(clippy::type_complexity)]
-fn init_morsel_distributor(
-    phase_tx_receivers: Vec<connector::Receiver<MorselOutput>>,
-) -> (
-    distributor_channel::Sender<(MorselSeq, DataFrame)>,
-    Vec<AbortOnDropHandle<PolarsResult<()>>>,
-) {
-    let (tx, dist_receivers) =
-        distributor_channel::<(MorselSeq, DataFrame)>(phase_tx_receivers.len(), 1);
-
-    let join_handles = phase_tx_receivers
-        .into_iter()
-        .zip(dist_receivers)
-        .map(|(mut phase_tx_receiver, mut morsel_rx)| {
-            AbortOnDropHandle::new(spawn(TaskPriority::Low, async move {
-                let Ok(mut morsel_output) = phase_tx_receiver.recv().await else {
-                    return Ok(());
-                };
-
-                let wait_group = WaitGroup::default();
-
-                'outer: loop {
-                    let Ok((morsel_seq, df)) = morsel_rx.recv().await else {
-                        return Ok(());
-                    };
-
-                    let mut morsel =
-                        Morsel::new(df, morsel_seq, morsel_output.source_token.clone());
-                    morsel.set_consume_token(wait_group.token());
-
-                    if morsel_output.port.send(morsel).await.is_err() {
-                        break 'outer;
-                    }
-
-                    wait_group.wait().await;
-
-                    if morsel_output.source_token.stop_requested() {
-                        morsel_output.outcome.stop();
-                        drop(morsel_output);
-
-                        let Ok(next_output) = phase_tx_receiver.recv().await else {
-                            break 'outer;
-                        };
-
-                        morsel_output = next_output;
-                    }
-                }
-
-                Ok(())
-            }))
-        })
-        .collect();
-
-    (tx, join_handles)
 }