--- a/crates/polars-stream/src/physical_plan/mod.rs
+++ b/crates/polars-stream/src/physical_plan/mod.rs
@@ -27,9 +27,11 @@ pub use fmt::visualize_plan;
 use polars_plan::prelude::{FileScanOptions, FileType};
 use polars_utils::arena::{Arena, Node};
 use polars_utils::pl_str::PlSmallStr;
+use polars_utils::slice_enum::Slice;
 use slotmap::{SecondaryMap, SlotMap};
 pub use to_graph::physical_plan_to_graph;
 
+use crate::nodes::io_sources::multi_file_reader::reader_interface::builder::FileReaderBuilder;
 use crate::nodes::io_sources::multi_scan::MultiscanRowRestriction;
 use crate::physical_plan::lower_expr::ExprCache;
 
@@ -213,12 +215,23 @@ pub enum PhysNodeKind {
         /// `allow_missing_columns == false`.
         file_schema: SchemaRef,
 
+        /// Final output schema of morsels being sent out of MultiScan.
+        output_schema: SchemaRef,
+
         /// Selection of `file_schema` columns should to be included in the output morsels.
         projection: Option<Bitmap>,
 
         row_restriction: Option<MultiscanRowRestriction>,
         predicate: Option<ExprIR>,
         row_index: Option<RowIndex>,
+
+        // Fields for new multiscan
+        // TODO: Remove `Option<>`
+        file_reader_builder: Option<Arc<dyn FileReaderBuilder>>,
+        /// Columns to project from the file.
+        projected_file_schema: SchemaRef,
+        cloud_options: Option<Arc<CloudOptions>>,
+        pre_slice: Option<Slice>,
     },
     FileScan {
         scan_source: ScanSource,