PR dispatches NDJSON to the updated `MultiFileReader`. It also contains generic IR lowering code for `FileReader` / `FileReaderBuilder` implementors.

For performance improvements, there is faster early-stopping on head/tail queries on a file list. Other queries otherwise perform the same.

#### Test on 100x 1M `jsonl` files:
Tested on AWS `c6gn.8xlarge`

* `head(10_000_000)`:
```
# Before (pypi 1.26.0)
81.13s user 8.29s system 1261% cpu 7.086 total
# After
63.15s user 6.73s system 1654% cpu 4.223 total
```
* `tail(10_000_000)`:
  * Note for more than 1 file we still do an extra pass to count rows.
```
# Before (pypi 1.26.0)
78.47s user 7.73s system 589% cpu 14.626 total
# After
70.48s user 6.42s system 652% cpu 11.786 total
```

@coastalwhite